move plugin liquidity provisiton to liquidity.svc
